% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_presentation.R
\name{create_presentation}
\alias{create_presentation}
\title{Create a New Quarto Presentation}
\usage{
create_presentation(path)
}
\arguments{
\item{path}{The directory where the new presentation should be created.}
}
\description{
This function copies a Quarto presentation template into a new directory.
}
