% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goterps_markdown.R
\name{goterps_markdown}
\alias{goterps_markdown}
\title{Create a New Quarto Presentation}
\usage{
goterps_markdown(path)
}
\arguments{
\item{path}{The directory where the new presentation should be created.}
}
\description{
This function copies a Quarto presentation template into a new directory.
}
