% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goterps_palette_c.R
\name{goterps_palette_c}
\alias{goterps_palette_c}
\title{Custom ggplot scale of colors for continuous variables.}
\usage{
goterps_palette_c(theme, n)
}
\arguments{
\item{theme}{A character string. The base color theme to use. Options are "yellow", "red", or "grey".}

\item{n}{An integer. Number of gradient colors to generate.}
}
\description{
This function applies a scale of UMD colors to continuous ggplot objects.
}
