% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goterps_palette_d.R
\name{goterps_palette_d}
\alias{goterps_palette_d}
\title{Custom ggplot scale of colors for continuous variables.}
\usage{
goterps_palette_d(n)
}
\arguments{
\item{n}{An integer. Number of gradient colors to generate.}

\item{theme}{A character string. The base color theme to use. Options are "yellow", "red", or "grey".}
}
\description{
This function applies a scale of UMD colors to continuous ggplot objects.
}
