% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goterps_presentation.R
\name{goterps_presentation}
\alias{goterps_presentation}
\title{Create and Duplicate Project Structure}
\usage{
goterps_presentation(path)
}
\arguments{
\item{path}{The path where the project structure should be created and files duplicated.}
}
\description{
This function creates a new project structure in a specified directory and duplicates all necessary files.
}
