% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goterps_palette_d.R
\name{goterps_palette_d}
\alias{goterps_palette_d}
\title{Custom ggplot scale of colors for discrete variable.}
\usage{
goterps_palette_d(n)
}
\description{
This function applies a scale of UMD colors to discrete ggplot objects.
}
